add new files and prod variant's and change 347 problem
